Create DateHelper Module
We decided to create a `DateHelper` Module and move all the formatted
date method. The reason for that is so every time we need to use them
we don't have to include or add_template_helper of ApplicationHelper,
which is not very secured.
