+++ app/helpers/application_helper.rb
@@ -1,16 +1,4 @@
 module ApplicationHelper
-  def date_from_string_or_date(obj)
-    obj.class == String ? Date.parse(obj) : obj
-  end
-
-  def format_date_of_birth(date)
-    I18n.l(date_from_string_or_date(date), format: :date_of_birth)
-  end
-
-  def format_date_of_visit(date)
-    I18n.l(date_from_string_or_date(date), format: :date_of_visit)
-  end
-
   def display_start_time(times)
     format_time_str(times.split('-')[0])
   end
