+++ app/mailers/prison_mailer.rb
@@ -3,10 +3,12 @@ class PrisonMailer < ActionMailer::Base
   include Autoresponder
   include Addresses
   include EnsureQuotedPrintable
-  include ApplicationHelper
+  include DateHelper
 
   after_action :do_not_send_to_prison, if: :smoke_test?
 
+  add_template_helper(DateHelper)
+  # needs to be removed after refactoring time
   add_template_helper(ApplicationHelper)
   add_template_helper(VisitHelper)
