+++ app/views/mailers_common/_booking_rejection.text.erb
@@ -7,7 +7,7 @@ You have tried to book a visit but the prisoner you want to visit has not got an
 We’re sorry but none of the dates and times you chose to visit <%= @visit.prisoner.first_name %> <%= initial ? @visit.prisoner.last_initial : @visit.prisoner.last_name %> at <%= @visit.prison_name %> were available.
 <% @visit.slots.each_with_index do |slot, index| %>
 <% if slot.date.present? %>
-<%= index + 1%>. <%= Date.parse(slot.date).strftime("%A %e %B") %>, <%= display_time_slot(slot.times) %>
+<%= index + 1%>. <%= format_date_of_visit(slot.date) %>, <%= display_time_slot(slot.times) %>
 <% end %>
 <% end %>
 
@@ -16,11 +16,11 @@ Please visit www.gov.uk/prison-visits to choose some alternative dates.
 We’re sorry, but the prisoner you want to visit has not got any visiting allowance left for the dates you’ve chosen.
 <% if @confirmation.no_pvo %>
 
-However, you can book a weekday visit with visiting allowance valid until <%= Date.parse(@confirmation.renew_pvo).strftime("%A %e %B") %>. The visit must be taken before the allowance expires.
+However, you can book a weekday visit with visiting allowance valid until <%= format_date_of_visit(@confirmation.renew_pvo) %>. The visit must be taken before the allowance expires.
 <% end %>
 <% if @confirmation.no_vo %>
 
-<%= @visit.prisoner.first_name %> <%= initial ? @visit.prisoner.last_initial : @visit.prisoner.last_name %> will have their full visiting allowance (VO) renewed on <%= Date.parse(@confirmation.renew_vo).strftime("%A %e %B") %>.
+<%= @visit.prisoner.first_name %> <%= initial ? @visit.prisoner.last_initial : @visit.prisoner.last_name %> will have their full visiting allowance (VO) renewed on <%= format_date_of_visit(@confirmation.renew_vo) %>.
 <% end %>
 <% when Confirmation::PRISONER_INCORRECT %>
 Your visit cannot take place as you haven’t given correct information for the prisoner. Eg, the prisoner’s name, number or date of birth is incorrect. Contact the prisoner to get up-to-date information.
