+++ spec/helpers/application_helper_spec.rb
@@ -11,22 +11,6 @@ RSpec.describe ApplicationHelper do
     end
   end
 
-  it "formats a date from a string" do
-    expect(helper.format_date_of_birth('2014-07-24')).to eq("24 July 2014")
-  end
-
-  it "formats a date from a date" do
-    expect(helper.format_date_of_birth(Date.parse('2014-07-24'))).to eq("24 July 2014")
-  end
-
-  it "formats a day from a string" do
-    expect(helper.format_date_of_visit('2014-07-24')).to eq("Thursday 24 July")
-  end
-
-  it "formats a day from a date" do
-    expect(helper.format_date_of_visit(Date.parse('2014-07-24'))).to eq("Thursday 24 July")
-  end
-
   it "formats a start time from a 24hr time string" do
     expect(helper.display_start_time('0945')).to eq("9:45am")
   end
